Further split code into modules.
